requirements.txt
asgiref==3.8.1
Django==5.1.1
djangorestframework==3.15.2
mysqlclient==2.2.4
pillow==10.4.0
sqlparse==0.5.1
tzdata==2024.2
